Work on the books reducer
