+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"/Users/FernandoBC/Documents/Microverse/React/book-store/src/reducers/index.js":"1","/Users/FernandoBC/Documents/Microverse/React/book-store/src/components/BooksList.js":"2","/Users/FernandoBC/Documents/Microverse/React/book-store/src/containers/BooksDetail.js":"3","/Users/FernandoBC/Documents/Microverse/React/book-store/src/reducers/books.js":"4","/Users/FernandoBC/Documents/Microverse/React/book-store/src/index.js":"5","/Users/FernandoBC/Documents/Microverse/React/book-store/src/App.js":"6","/Users/FernandoBC/Documents/Microverse/React/book-store/src/components/BooksForm.js":"7","/Users/FernandoBC/Documents/Microverse/React/book-store/src/components/Book.js":"8"},{"size":166,"mtime":1612374036169,"results":"9","hashOfConfig":"10"},{"size":309,"mtime":1612374767375,"results":"11","hashOfConfig":"10"},{"size":506,"mtime":1612381590604,"results":"12","hashOfConfig":"10"},{"size":439,"mtime":1612380749948,"results":"13","hashOfConfig":"10"},{"size":358,"mtime":1612373680130,"results":"14","hashOfConfig":"10"},{"size":213,"mtime":1612373680112,"results":"15","hashOfConfig":"10"},{"size":410,"mtime":1612382657658,"results":"16","hashOfConfig":"10"},{"size":416,"mtime":1612381517938,"results":"17","hashOfConfig":"10"},{"filePath":"18","messages":"19","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"1v7fwdc",{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},{"filePath":"29","messages":"30","errorCount":0,"warningCount":1,"fixableErrorCount":0,"fixableWarningCount":0,"source":"31","usedDeprecatedRules":"20"},{"filePath":"32","messages":"33","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"34","messages":"35","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"20"},"/Users/FernandoBC/Documents/Microverse/React/book-store/src/reducers/index.js",[],["36","37","38","39","40","41"],"/Users/FernandoBC/Documents/Microverse/React/book-store/src/components/BooksList.js",[],"/Users/FernandoBC/Documents/Microverse/React/book-store/src/containers/BooksDetail.js",[],"/Users/FernandoBC/Documents/Microverse/React/book-store/src/reducers/books.js",[],"/Users/FernandoBC/Documents/Microverse/React/book-store/src/index.js",[],"/Users/FernandoBC/Documents/Microverse/React/book-store/src/App.js",["42"],"import BooksList from './components/BooksList';\nimport BooksForm from './components/BooksForm';\n\nconst App = () => (\n  <div>\n    <BooksList />\n    <BooksForm />\n  </div>\n);\n\nconsole.log(App);\n\nexport default App;\n","/Users/FernandoBC/Documents/Microverse/React/book-store/src/components/BooksForm.js",[],"/Users/FernandoBC/Documents/Microverse/React/book-store/src/components/Book.js",[],{"ruleId":"43","replacedBy":"44"},{"ruleId":"45","replacedBy":"46"},{"ruleId":"47","replacedBy":"48"},{"ruleId":"49","replacedBy":"50"},{"ruleId":"51","replacedBy":"52"},{"ruleId":"53","replacedBy":"54"},{"ruleId":"55","severity":1,"message":"56","line":11,"column":1,"nodeType":"57","messageId":"58","endLine":11,"endColumn":12},"lines-around-directive",["59"],"no-spaced-func",["60"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"no-console","Unexpected console statement.","MemberExpression","unexpected","padding-line-between-statements","func-call-spacing"]
+[{"/home/tadeuasarro/Documents/book-store/src/index.js":"1","/home/tadeuasarro/Documents/book-store/src/App.js":"2","/home/tadeuasarro/Documents/book-store/src/reducers/index.js":"3","/home/tadeuasarro/Documents/book-store/src/components/BooksList.js":"4","/home/tadeuasarro/Documents/book-store/src/components/BooksForm.js":"5","/home/tadeuasarro/Documents/book-store/src/containers/BooksDetail.js":"6","/home/tadeuasarro/Documents/book-store/src/components/Book.js":"7","/home/tadeuasarro/Documents/book-store/src/reducers/booksData.js":"8","/home/tadeuasarro/Documents/book-store/src/reducers/books.js":"9","/home/tadeuasarro/Documents/book-store/src/actions/constants.js":"10"},{"size":358,"mtime":1612383492896,"results":"11","hashOfConfig":"12"},{"size":194,"mtime":1612384745932,"results":"13","hashOfConfig":"12"},{"size":228,"mtime":1612384898809,"results":"14","hashOfConfig":"12"},{"size":309,"mtime":1612383492896,"results":"15","hashOfConfig":"12"},{"size":410,"mtime":1612383492896,"results":"16","hashOfConfig":"12"},{"size":496,"mtime":1612385364106,"results":"17","hashOfConfig":"12"},{"size":416,"mtime":1612383492896,"results":"18","hashOfConfig":"12"},{"size":549,"mtime":1612385463694,"results":"19","hashOfConfig":"12"},{"size":352,"mtime":1612385173031,"results":"20","hashOfConfig":"12"},{"size":84,"mtime":1612383492896,"results":"21","hashOfConfig":"12"},{"filePath":"22","messages":"23","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},"5ltmzh",{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"31","messages":"32","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"33","messages":"34","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"35","messages":"36","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"24"},{"filePath":"37","messages":"38","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"39","messages":"40","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"41","messages":"42","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"/home/tadeuasarro/Documents/book-store/src/index.js",[],["43","44","45","46","47","48"],"/home/tadeuasarro/Documents/book-store/src/App.js",[],"/home/tadeuasarro/Documents/book-store/src/reducers/index.js",[],"/home/tadeuasarro/Documents/book-store/src/components/BooksList.js",[],"/home/tadeuasarro/Documents/book-store/src/components/BooksForm.js",[],"/home/tadeuasarro/Documents/book-store/src/containers/BooksDetail.js",[],"/home/tadeuasarro/Documents/book-store/src/components/Book.js",[],"/home/tadeuasarro/Documents/book-store/src/reducers/booksData.js",[],"/home/tadeuasarro/Documents/book-store/src/reducers/books.js",[],"/home/tadeuasarro/Documents/book-store/src/actions/constants.js",[],{"ruleId":"49","replacedBy":"50"},{"ruleId":"51","replacedBy":"52"},{"ruleId":"53","replacedBy":"54"},{"ruleId":"55","replacedBy":"56"},{"ruleId":"57","replacedBy":"58"},{"ruleId":"59","replacedBy":"60"},"lines-around-directive",["61"],"no-spaced-func",["62"],"global-require",[],"no-buffer-constructor",[],"no-new-require",[],"no-path-concat",[],"padding-line-between-statements","func-call-spacing"]
