+++ src/actions/index.js
@@ -1,9 +1,11 @@
 import { CREATE_BOOK, REMOVE_BOOK } from './constants';
 
-export const addBook = () => ({
+export const addBook = book => ({
   type: CREATE_BOOK,
+  book: book,
 });
 
-export const deleteBook = () => ({
+export const deleteBook = book => ({
   type: REMOVE_BOOK,
+  book: book
 });
